PCB layout updates
I updated the PCB with the components added to the schematic. I dimensioned a rectangle around the board that is 2x1.5 in and placed the 8 pin connecters with respect to the corners of the rectangle. I added the relevant dimensions. Board layout based on our discussion Friday. I remember we were having some discussion on the 4 connectors in each corner of the board, those have not been added yet.