Ran DRC
There were only warnings no errors. One useless via was placed, I got rid of it. All other errors referred to the Silkscreen overlapping or being cut off by a component/via. I believe this just refers to the writing printed on the board so I am assuming this is not important.